Load example phylospatial data on California mosses

This function returns the moss phylospatial data set, with options for raster vs polygon formatting and for probability vs binary community data type.

Usage

moss_data(
  data_type = c("probability", "binary"),
  spatial_type = c("raster", "polygons")
)

Arguments

data_type

Either "probability" or "binary".

spatial_type

Either "raster" or "polygons".

Value

A phylospatial object with a phylogeny and modeled distributions of 443 species of mosses, representing a spatially coarser version of data published by Kling et al. (2024).

References

Kling, Gonzalez-Ramirez, et al. (2024) bioRxiv, https://doi.org/10.1101/2024.12.16.628580.

Examples

moss_data("prob", "poly")
#> `phylospatial` object
#>   - 884 lineages across 171 sites
#>   - community data type: probability 
#>   - spatial data class: sf 
#>   - dissimilarity data: none 

# \donttest{
moss_data("bin", "rast")
#> `phylospatial` object
#>   - 884 lineages across 288 sites
#>   - community data type: binary 
#>   - spatial data class: SpatRaster 
#>   - dissimilarity data: none 
# }