Binary randomization tests including CANAPE

This function is a wrapper around the cpr_rand_test function from the canaper package. In keeping with CANAPE, it requires a binary community matrix and uses a hard significance threshold (for an alternative that utilizes continuous community values and retains a significance gradient, see ps_rand).

Usage

ps_canape(ps, null_model = "curveball", spatial = TRUE, ...)

Arguments

ps

phylospatial object

null_model

see canaper::cpr_rand_test

spatial

Logical: should the function return a spatial object (TRUE, default) or a vector (FALSE).

...

further arguments passed to canaper::cpr_rand_test

Value

A matrix or raster stack with a column or layer (respectively) for each metric.

Details

This function runs canaper::cpr_rand_test; see the help for that function for details.

It also runs canaper::cpr_classify_endem on the result, and includes the resulting classification as an additional variable, 'endem_type', in the output. 'endem_type' values 0-4 correspond to not-significant, neo, paleo, mixed, and super endemesim, respectively.

References

Mishler, B. D., Knerr, N., González-Orozco, C. E., Thornhill, A. H., Laffan, S. W., & Miller, J. T. (2014). Phylogenetic measures of biodiversity and neo-and paleo-endemism in Australian Acacia. Nature Communications, 5(1), 4473.

Nitta, J. H., Laffan, S. W., Mishler, B. D., & Iwasaki, W. (2023). canaper: categorical analysis of neo‐and paleo‐endemism in R. Ecography, 2023(9), e06638.

Examples

ps <- ps_simulate(data_type = "binary")
ps_canape(ps)
#> class       : SpatRaster 
#> dimensions  : 20, 20, 55  (nrow, ncol, nlyr)
#> resolution  : 1, 1  (x, y)
#> extent      : 0, 20, 0, 20  (xmin, xmax, ymin, ymax)
#> coord. ref. :  
#> source(s)   : memory
#> varnames    : pd_obs 
#>               pd_rand_mean 
#>               pd_rand_sd 
#>               ...
#> names       :    pd_obs, pd_rand_mean, pd_rand_sd,  pd_obs_z, pd_ob~upper, pd_ob~lower, ... 
#> min values  : 0.1419149,    0.2642274, 0.02773481, -3.327005,           0,           0, ... 
#> max values  : 0.8234926,    0.8158520, 0.07372272,  1.733122,         100,         100, ...