Expand occupied-only results to full spatial extent

Takes a matrix or vector computed on occupied sites only and expands it back to the full site grid, inserting NA for unoccupied sites. This is useful when performing custom analyses on ps$comm (which contains only occupied sites) and mapping the results back to the full raster or spatial object.

Usage

ps_expand(ps, x, spatial = FALSE)

Arguments

ps

A phylospatial object.

x

A matrix (or vector) with nrow(ps$comm) rows (i.e. one row per occupied site).

spatial

Logical: if TRUE, convert the expanded result to a spatial object using ps$spatial. Default is FALSE.

Value

If spatial = FALSE, a matrix with ps$n_sites rows and NA for unoccupied sites. If spatial = TRUE, a SpatRaster or sf object.

Examples

ps <- ps_simulate()

# custom analysis on the occupied-only community matrix
site_totals <- matrix(rowSums(ps$comm), ncol = 1)
colnames(site_totals) <- "total"

# expand to full extent as a matrix
ps_expand(ps, site_totals)
#>            total
#>   [1,] 0.2286463
#>   [2,] 0.2548852
#>   [3,] 0.2775980
#>   [4,] 0.2966795
#>   [5,] 0.3120737
#>   [6,] 0.3237788
#>   [7,] 0.3318490
#>   [8,] 0.3092227
#>   [9,] 0.2831033
#>  [10,] 0.2536352
#>  [11,] 0.3249341
#>  [12,] 0.7876618
#>  [13,] 1.2533376
#>  [14,] 1.7378916
#>  [15,] 2.2224457
#>  [16,] 2.7069997
#>  [17,] 3.1915537
#>  [18,] 3.6761078
#>  [19,] 4.1606618
#>  [20,] 4.6452158
#>  [21,] 0.2775810
#>  [22,] 0.3043922
#>  [23,] 0.3273404
#>  [24,] 0.3461945
#>  [25,] 0.3606177
#>  [26,] 0.3696655
#>  [27,] 0.3693179
#>  [28,] 0.3540050
#>  [29,] 0.3294966
#>  [30,] 0.3000438
#>  [31,] 0.3431368
#>  [32,] 0.8015159
#>  [33,] 1.2559729
#>  [34,] 1.7063205
#>  [35,] 2.1722408
#>  [36,] 2.6448864
#>  [37,] 3.1099951
#>  [38,] 3.5578382
#>  [39,] 3.9536658
#>  [40,] 4.1463387
#>  [41,] 0.3240739
#>  [42,] 0.3512274
#>  [43,] 0.3740118
#>  [44,] 0.3919874
#>  [45,] 0.4044815
#>  [46,] 0.4102071
#>  [47,] 0.4069390
#>  [48,] 0.3932502
#>  [49,] 0.3708482
#>  [50,] 0.3422363
#>  [51,] 0.3089447
#>  [52,] 0.7514169
#>  [53,] 1.1924818
#>  [54,] 1.6254321
#>  [55,] 2.0478050
#>  [56,] 2.4686837
#>  [57,] 2.8870472
#>  [58,] 3.2607758
#>  [59,] 3.5423874
#>  [60,] 3.6464114
#>  [61,] 0.3683265
#>  [62,] 0.3957702
#>  [63,] 0.4182914
#>  [64,] 0.4353044
#>  [65,] 0.4460569
#>  [66,] 0.4495451
#>  [67,] 0.4446875
#>  [68,] 0.4310641
#>  [69,] 0.4094757
#>  [70,] 0.3813565
#>  [71,] 0.3480003
#>  [72,] 0.6415479
#>  [73,] 1.0629065
#>  [74,] 1.4709131
#>  [75,] 1.8611765
#>  [76,] 2.3319611
#>  [77,] 2.7516666
#>  [78,] 3.0766609
#>  [79,] 3.2501869
#>  [80,] 3.2379922
#>  [81,] 0.4105751
#>  [82,] 0.4384433
#>  [83,] 0.4608619
#>  [84,] 0.4771683
#>  [85,] 0.4866559
#>  [86,] 0.4886194
#>  [87,] 0.4825287
#>  [88,] 0.4683272
#>  [89,] 0.4465857
#>  [90,] 0.4182914
#>  [91,] 0.3845025
#>  [92,] 0.4778609
#>  [93,] 0.8752481
#>  [94,] 1.2540543
#>  [95,] 1.7874852
#>  [96,] 2.2701812
#>  [97,] 2.6979309
#>  [98,] 3.0543276
#>  [99,] 3.1380615
#> [100,] 3.0270348
#> [101,] 0.4510088
#> [102,] 0.4795906
#> [103,] 0.5022339
#> [104,] 0.5182138
#> [105,] 0.5268679
#> [106,] 0.5558947
#> [107,] 0.6146616
#> [108,] 0.6410732
#> [109,] 0.6327297
#> [110,] 0.5893592
#> [111,] 0.5130253
#> [112,] 0.4075816
#> [113,] 0.6381686
#> [114,] 1.1109603
#> [115,] 1.6745301
#> [116,] 2.4861678
#> [117,] 3.1287853
#> [118,] 3.4668297
#> [119,] 3.4759494
#> [120,] 3.1719128
#> [121,] 0.4897113
#> [122,] 0.5194197
#> [123,] 0.5427235
#> [124,] 0.5587984
#> [125,] 0.6476914
#> [126,] 0.7460226
#> [127,] 0.8142111
#> [128,] 0.8468523
#> [129,] 0.8398713
#> [130,] 0.7922617
#> [131,] 0.7068462
#> [132,] 0.5892571
#> [133,] 0.4460136
#> [134,] 0.9153061
#> [135,] 1.8179252
#> [136,] 2.7664303
#> [137,] 3.5178627
#> [138,] 3.9203618
#> [139,] 3.8164927
#> [140,] 3.3757719
#> [141,] 0.5266180
#> [142,] 0.5579790
#> [143,] 0.5824783
#> [144,] 0.6798355
#> [145,] 0.8162691
#> [146,] 0.9282623
#> [147,] 1.0090844
#> [148,] 1.0506900
#> [149,] 1.0458332
#> [150,] 0.9920407
#> [151,] 0.8938339
#> [152,] 0.7602628
#> [153,] 0.6007194
#> [154,] 0.7657757
#> [155,] 1.8183623
#> [156,] 2.8508136
#> [157,] 3.7504259
#> [158,] 4.3917585
#> [159,] 4.0644222
#> [160,] 3.4476191
#> [161,] 0.5614712
#> [162,] 0.5957488
#> [163,] 0.6497047
#> [164,] 0.8184441
#> [165,] 0.9695590
#> [166,] 1.0978887
#> [167,] 1.1956482
#> [168,] 1.2510366
#> [169,] 1.2502492
#> [170,] 1.1870661
#> [171,] 1.0701659
#> [172,] 0.9155631
#> [173,] 0.7367404
#> [174,] 0.9654586
#> [175,] 1.6896804
#> [176,] 2.6209331
#> [177,] 3.5696153
#> [178,] 4.3197526
#> [179,] 3.9731217
#> [180,] 3.2652876
#> [181,] 0.6757938
#> [182,] 0.7214855
#> [183,] 0.8349041
#> [184,] 0.9894873
#> [185,] 1.1158684
#> [186,] 1.2467907
#> [187,] 1.3661991
#> [188,] 1.4440277
#> [189,] 1.4525054
#> [190,] 1.3733447
#> [191,] 1.2278504
#> [192,] 1.0466070
#> [193,] 0.8468246
#> [194,] 0.9342495
#> [195,] 1.6054462
#> [196,] 2.3050245
#> [197,] 2.9922878
#> [198,] 3.4287092
#> [199,] 3.4913958
#> [200,] 2.7314399
#> [201,] 0.7928112
#> [202,] 0.8446750
#> [203,] 0.9994916
#> [204,] 1.1573711
#> [205,] 1.2870327
#> [206,] 1.3912473
#> [207,] 1.5037895
#> [208,] 1.6461810
#> [209,] 1.7202106
#> [210,] 1.6208133
#> [211,] 1.4198582
#> [212,] 1.1703912
#> [213,] 0.9218346
#> [214,] 0.8113624
#> [215,] 1.2729070
#> [216,] 1.8287251
#> [217,] 2.3748194
#> [218,] 2.6818630
#> [219,] 2.6138696
#> [220,] 2.1026809
#> [221,] 0.9035973
#> [222,] 0.9639265
#> [223,] 1.1329266
#> [224,] 1.2882264
#> [225,] 1.4129439
#> [226,] 1.5119693
#> [227,] 1.6757332
#> [228,] 1.8707109
#> [229,] 2.0453651
#> [230,] 1.8474084
#> [231,] 1.6119926
#> [232,] 1.3409642
#> [233,] 1.0399680
#> [234,] 0.7554566
#> [235,] 0.7893518
#> [236,] 1.1745827
#> [237,] 1.6021656
#> [238,] 1.8093514
#> [239,] 1.7259728
#> [240,] 1.3361634
#> [241,] 1.0043164
#> [242,] 1.0767895
#> [243,] 1.2310445
#> [244,] 1.3764617
#> [245,] 1.4873012
#> [246,] 1.6496878
#> [247,] 1.8067332
#> [248,] 1.9424622
#> [249,] 2.0252577
#> [250,] 1.9282627
#> [251,] 1.7162341
#> [252,] 1.4440073
#> [253,] 1.1340027
#> [254,] 0.7990660
#> [255,] 0.6251421
#> [256,] 0.6531823
#> [257,] 0.8085349
#> [258,] 0.9213012
#> [259,] 0.8629043
#> [260,] 0.7306129
#> [261,] 1.0860163
#> [262,] 1.1793496
#> [263,] 1.2844860
#> [264,] 1.4126838
#> [265,] 1.5383414
#> [266,] 1.7112204
#> [267,] 1.8402256
#> [268,] 1.9262584
#> [269,] 1.9959661
#> [270,] 1.9365753
#> [271,] 1.7510511
#> [272,] 1.4817972
#> [273,] 1.1670367
#> [274,] 0.8265684
#> [275,] 0.6253188
#> [276,] 0.8245145
#> [277,] 0.9941720
#> [278,] 0.9461398
#> [279,] 0.6936067
#> [280,] 0.4544733
#> [281,] 1.1187515
#> [282,] 1.2669219
#> [283,] 1.2601533
#> [284,] 1.3704163
#> [285,] 1.5474634
#> [286,] 1.6763326
#> [287,] 1.7839889
#> [288,] 1.8556177
#> [289,] 1.9446719
#> [290,] 1.9225899
#> [291,] 1.7354579
#> [292,] 1.4553884
#> [293,] 1.1363365
#> [294,] 1.1005993
#> [295,] 1.1978603
#> [296,] 1.3546148
#> [297,] 1.4372573
#> [298,] 1.3428099
#> [299,] 1.0077632
#> [300,] 0.4854638
#> [301,] 1.0633598
#> [302,] 1.1563571
#> [303,] 1.1718696
#> [304,] 1.2214752
#> [305,] 1.3743244
#> [306,] 1.5252624
#> [307,] 1.6327939
#> [308,] 1.7118817
#> [309,] 1.8251879
#> [310,] 1.8974658
#> [311,] 1.6402200
#> [312,] 1.3509861
#> [313,] 1.1840382
#> [314,] 1.4427660
#> [315,] 1.7129308
#> [316,] 1.9987977
#> [317,] 2.1116718
#> [318,] 1.8920334
#> [319,] 1.4214678
#> [320,] 0.8207698
#> [321,] 0.9591061
#> [322,] 1.0306350
#> [323,] 1.0541941
#> [324,] 1.0349343
#> [325,] 1.1317684
#> [326,] 1.2843543
#> [327,] 1.3914463
#> [328,] 1.4822016
#> [329,] 1.5728484
#> [330,] 1.5677369
#> [331,] 1.4071518
#> [332,] 1.1570311
#> [333,] 1.2497168
#> [334,] 1.6409834
#> [335,] 2.1046109
#> [336,] 2.5703230
#> [337,] 2.8092404
#> [338,] 2.4646535
#> [339,] 1.8329082
#> [340,] 1.1275990
#> [341,] 0.8471248
#> [342,] 0.9023609
#> [343,] 0.9218233
#> [344,] 0.9055088
#> [345,] 0.8885212
#> [346,] 0.9782372
#> [347,] 1.0774858
#> [348,] 1.1856448
#> [349,] 1.2476625
#> [350,] 1.2253402
#> [351,] 1.1031706
#> [352,] 0.9013561
#> [353,] 1.1779149
#> [354,] 1.6920024
#> [355,] 2.2224769
#> [356,] 2.8501544
#> [357,] 3.5025309
#> [358,] 2.7462364
#> [359,] 1.9897378
#> [360,] 1.2330742
#> [361,] 0.7251052
#> [362,] 0.7714173
#> [363,] 0.7885688
#> [364,] 0.7757229
#> [365,] 0.7346128
#> [366,] 0.7598323
#> [367,] 0.8259141
#> [368,] 0.8953156
#> [369,] 0.9280915
#> [370,] 0.9129954
#> [371,] 0.8448328
#> [372,] 0.7312752
#> [373,] 0.9800107
#> [374,] 1.4902564
#> [375,] 2.0135307
#> [376,] 2.5422338
#> [377,] 2.7833354
#> [378,] 2.4406267
#> [379,] 1.8105121
#> [380,] 1.1066308
#> [381,] 0.5969324
#> [382,] 0.6365315
#> [383,] 0.6514629
#> [384,] 0.6409315
#> [385,] 0.6059487
#> [386,] 0.5847495
#> [387,] 0.6608864
#> [388,] 0.7179383
#> [389,] 0.7408011
#> [390,] 0.7238949
#> [391,] 0.6654235
#> [392,] 0.5694800
#> [393,] 0.6001931
#> [394,] 1.0745594
#> [395,] 1.5744234
#> [396,] 1.9432105
#> [397,] 2.0603315
#> [398,] 1.8443583
#> [399,] 1.3769845
#> [400,] 0.7790886

# expand and convert to spatial
ps_expand(ps, site_totals, spatial = TRUE)
#> class       : SpatRaster 
#> size        : 20, 20, 1  (nrow, ncol, nlyr)
#> resolution  : 1, 1  (x, y)
#> extent      : 0, 20, 0, 20  (xmin, xmax, ymin, ymax)
#> coord. ref. :  
#> source(s)   : memory
#> varname     : total 
#> name        :     total 
#> min value   : 0.2286463 
#> max value   : 4.6452158