Get phylospatial community data

Get phylospatial community data

Usage

ps_get_comm(ps, tips_only = TRUE, spatial = TRUE)

Arguments

ps

phylospatial object.

tips_only

Logical indicating whether only the terminal taxa (TRUE, the default) or all taxa (FALSE) should be returned.

spatial

Logical indicating whether a spatial (SpatRaster or sf) object should be returned. Default is TRUE; if FALSE, a matrix is returned.

Value

Either a SpatRaster with a layer for every taxon, or an sf data frame with a variable for every taxon, depending on which data type was used to create ps.

Examples

ps <- ps_simulate()

# the defaults return a spatial object of terminal taxa distributions:
ps_get_comm(ps)
#> class       : SpatRaster 
#> dimensions  : 20, 20, 10  (nrow, ncol, nlyr)
#> resolution  : 1, 1  (x, y)
#> extent      : 0, 20, 0, 20  (xmin, xmax, ymin, ymax)
#> coord. ref. :  
#> source(s)   : memory
#> varnames    : t3 
#>               t2 
#>               t5 
#>               ...
#> names       :        t3,        t2,        t5,        t9,        t1,        t8, ... 
#> min values  : 0.0000000, 0.0000000, 0.0000000, 0.0000000, 0.0000000, 0.0000000, ... 
#> max values  : 0.2514489, 0.6154435, 0.8641071, 0.7620122, 0.9165357, 0.1465094, ... 

# get distributions for all taxa, as a matrix
pcomm <- ps_get_comm(ps, tips_only = FALSE, spatial = FALSE)