Convert a site-by-variable matrix into a SpatRaster or sf object

Convert a site-by-variable matrix into a SpatRaster or sf object

Usage

to_spatial(m, template)

Arguments

m

Matrix.

template

SpatRaster layer with number of cells equal to the number of rows in m, or sf data frame with same number of rows as m.

Value

SpatRaster with a layer for every column in m, or sf data frame with a variable for every column in m, depending on the data type of template.

Examples

# example using `sf` data set `moss`:
to_spatial(moss$comm[, 1:5], moss$spatial)
#> Simple feature collection with 171 features and 5 fields
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -405000 ymin: -664641 xmax: 585000 ymax: 495833
#> Projected CRS: unnamed
#> First 10 features:
#>         clade1 Sphagnum_teres      clade2 Sphagnum_squarrosum     clade3
#> 1  0.271485552   2.012270e-08 0.271485537        9.046410e-07 0.27148488
#> 2  0.029584329   1.896249e-27 0.029584329        1.633291e-05 0.02956848
#> 3  0.407162761   3.776328e-02 0.383896682        1.884237e-18 0.38389668
#> 4  0.185452662   1.466808e-04 0.185333166        1.909400e-07 0.18533301
#> 5  0.777731999   1.241481e-14 0.777731999        3.489658e-04 0.77765441
#> 6  0.006214871   3.463186e-27 0.006214871        1.267038e-03 0.00495411
#> 7  0.451858304   4.797141e-02 0.424238197        8.726398e-18 0.42423820
#> 8  0.669838232   3.318939e-02 0.658504193        2.698501e-10 0.65850419
#> 9  0.815675026   5.281141e-08 0.815675016        1.311591e-03 0.81543294
#> 10 0.238225139   6.056255e-19 0.238225139        1.165790e-02 0.22923970
#>                                 x
#> 1  POLYGON ((-375000 270666.4,...
#> 2  POLYGON ((-375000 374589.5,...
#> 3  POLYGON ((-345000 114781.8,...
#> 4  POLYGON ((-345000 218704.9,...
#> 5  POLYGON ((-345000 322627.9,...
#> 6  POLYGON ((-345000 426551, -...
#> 7  POLYGON ((-315000 62820.32,...
#> 8  POLYGON ((-315000 166743.4,...
#> 9  POLYGON ((-315000 270666.4,...
#> 10 POLYGON ((-315000 374589.5,...

# and using a `SpatRaster`:
to_spatial(moss_data()$comm[, 1:5], moss_data()$spatial)
#> class       : SpatRaster 
#> dimensions  : 18, 16, 5  (nrow, ncol, nlyr)
#> resolution  : 60000, 60000  (x, y)
#> extent      : -375000, 585000, -630000, 450000  (xmin, xmax, ymin, ymax)
#> coord. ref. : +proj=aea +lat_0=0 +lon_0=-120 +lat_1=34 +lat_2=40.5 +x_0=0 +y_0=-4000000 +datum=NAD83 +units=m +no_defs 
#> source(s)   : memory
#> varnames    : clade1 
#>               Sphagnum_teres 
#>               clade2 
#>               ...
#> names       :    clade1, Sphagnum_teres,    clade2, Sphagn~rrosum,    clade3 
#> min values  : 0.0000000,      0.0000000, 0.0000000,     0.0000000, 0.0000000 
#> max values  : 0.9318513,      0.2934258, 0.9318242,     0.3475139, 0.9198755