Convert a site-by-variable matrix into a SpatRaster or sf object

Convert a site-by-variable matrix into a SpatRaster or sf object

Usage

to_spatial(m, template)

Arguments

m

Matrix or vector.

template

SpatRaster layer with number of cells equal to the number of rows in m, or sf data frame with same number of rows as m.

Value

SpatRaster with a layer for every column in m, or sf data frame with a variable for every column in m, depending on the data type of template.

Examples

# example using `sf` data set `moss_hex`:
to_spatial(moss("polygon")$comm[, 1:5], moss("polygon")$spatial)
#> Simple feature collection with 527 features and 5 fields
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -375000 ymin: -630000 xmax: 555000 ymax: 450000
#> Projected CRS: unnamed
#> First 10 features:
#>          clade1 Sphagnum_teres       clade2 Sphagnum_squarrosum       clade3
#> 1  0.0153801784   1.832663e-29 0.0153801784        4.061571e-05 0.0153401858
#> 2  0.0270682060   2.796592e-25 0.0270682060        2.341211e-03 0.0247850219
#> 3  0.0581547094   1.398105e-21 0.0581547094        4.884653e-02 0.0097861997
#> 4  0.2086119861   2.853711e-18 0.2086119861        2.016190e-01 0.0087590150
#> 5  0.2530552111   1.080983e-15 0.2530552111        2.465081e-01 0.0086890374
#> 6  0.0584494860   1.026890e-13 0.0584494860        5.655514e-02 0.0020079025
#> 7  0.0170809968   4.367453e-12 0.0170809968        1.435780e-02 0.0027628668
#> 8  0.0004397862   4.552663e-11 0.0004397862        1.957388e-04 0.0002440952
#> 9  0.0001781171   7.732686e-11 0.0001781170        1.256262e-06 0.0001768609
#> 10 0.0127500322   7.544112e-11 0.0127500322        4.675405e-09 0.0127500275
#>                                 x
#> 1  POLYGON ((-375000 420000, -...
#> 2  POLYGON ((-345000 420000, -...
#> 3  POLYGON ((-315000 420000, -...
#> 4  POLYGON ((-285000 420000, -...
#> 5  POLYGON ((-255000 420000, -...
#> 6  POLYGON ((-225000 420000, -...
#> 7  POLYGON ((-195000 420000, -...
#> 8  POLYGON ((-165000 420000, -...
#> 9  POLYGON ((-135000 420000, -...
#> 10 POLYGON ((-105000 420000, -...

# and using a `SpatRaster` data set `moss`:
to_spatial(moss("polygon")$comm[, 1:5], moss("polygon")$spatial)
#> Simple feature collection with 527 features and 5 fields
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -375000 ymin: -630000 xmax: 555000 ymax: 450000
#> Projected CRS: unnamed
#> First 10 features:
#>          clade1 Sphagnum_teres       clade2 Sphagnum_squarrosum       clade3
#> 1  0.0153801784   1.832663e-29 0.0153801784        4.061571e-05 0.0153401858
#> 2  0.0270682060   2.796592e-25 0.0270682060        2.341211e-03 0.0247850219
#> 3  0.0581547094   1.398105e-21 0.0581547094        4.884653e-02 0.0097861997
#> 4  0.2086119861   2.853711e-18 0.2086119861        2.016190e-01 0.0087590150
#> 5  0.2530552111   1.080983e-15 0.2530552111        2.465081e-01 0.0086890374
#> 6  0.0584494860   1.026890e-13 0.0584494860        5.655514e-02 0.0020079025
#> 7  0.0170809968   4.367453e-12 0.0170809968        1.435780e-02 0.0027628668
#> 8  0.0004397862   4.552663e-11 0.0004397862        1.957388e-04 0.0002440952
#> 9  0.0001781171   7.732686e-11 0.0001781170        1.256262e-06 0.0001768609
#> 10 0.0127500322   7.544112e-11 0.0127500322        4.675405e-09 0.0127500275
#>                                 x
#> 1  POLYGON ((-375000 420000, -...
#> 2  POLYGON ((-345000 420000, -...
#> 3  POLYGON ((-315000 420000, -...
#> 4  POLYGON ((-285000 420000, -...
#> 5  POLYGON ((-255000 420000, -...
#> 6  POLYGON ((-225000 420000, -...
#> 7  POLYGON ((-195000 420000, -...
#> 8  POLYGON ((-165000 420000, -...
#> 9  POLYGON ((-135000 420000, -...
#> 10 POLYGON ((-105000 420000, -...