Convert a site-by-variable matrix into a SpatRaster or sf object

Convert a site-by-variable matrix into a SpatRaster or sf object

Usage

to_spatial(m, template)

Arguments

m

Matrix or vector.

template

SpatRaster layer with number of cells equal to the number of rows in m, or sf data frame with same number of rows as m.

Value

SpatRaster with a layer for every column in m, or sf data frame with a variable for every column in m, depending on the data type of template.

Examples

# example using `sf` data set `moss`:
to_spatial(moss$comm[, 1:5], moss$spatial)
#> Simple feature collection with 171 features and 5 fields
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -405000 ymin: -664641 xmax: 585000 ymax: 495833
#> Projected CRS: unnamed
#> First 10 features:
#>         clade1 Sphagnum_teres      clade2 Sphagnum_squarrosum      clade3
#> 1  0.310208217   2.347649e-08 0.310208201        1.055415e-06 0.310207473
#> 2  0.100672263   6.636870e-27 0.100672263        5.716519e-05 0.100620850
#> 3  0.978502146   2.265797e-01 0.972204179        1.130542e-17 0.972204179
#> 4  0.198646193   1.579639e-04 0.198519588        2.056277e-07 0.198519424
#> 5  0.895501523   1.738073e-14 0.895501523        4.885521e-04 0.895450445
#> 6  0.008279633   4.617581e-27 0.008279633        1.689384e-03 0.006601401
#> 7  0.627060745   7.675426e-02 0.596056349        1.396224e-17 0.596056349
#> 8  0.732783698   3.872096e-02 0.722020049        3.148251e-10 0.722020048
#> 9  0.815675026   5.281141e-08 0.815675016        1.311591e-03 0.815432939
#> 10 0.238225139   6.056255e-19 0.238225139        1.165790e-02 0.229239696
#>                                 x
#> 1  POLYGON ((-375000 270666.4,...
#> 2  POLYGON ((-375000 374589.5,...
#> 3  POLYGON ((-345000 114781.8,...
#> 4  POLYGON ((-345000 218704.9,...
#> 5  POLYGON ((-345000 322627.9,...
#> 6  POLYGON ((-345000 426551, -...
#> 7  POLYGON ((-315000 62820.32,...
#> 8  POLYGON ((-315000 166743.4,...
#> 9  POLYGON ((-315000 270666.4,...
#> 10 POLYGON ((-315000 374589.5,...

# and using a `SpatRaster`:
to_spatial(moss_data()$comm[, 1:5], moss_data()$spatial)
#> class       : SpatRaster 
#> dimensions  : 18, 16, 5  (nrow, ncol, nlyr)
#> resolution  : 60000, 60000  (x, y)
#> extent      : -375000, 585000, -630000, 450000  (xmin, xmax, ymin, ymax)
#> coord. ref. : +proj=aea +lat_0=0 +lon_0=-120 +lat_1=34 +lat_2=40.5 +x_0=0 +y_0=-4000000 +datum=NAD83 +units=m +no_defs 
#> source(s)   : memory
#> varnames    : clade1 
#>               Sphagnum_teres 
#>               clade2 
#>               ...
#> names       :    clade1, Sphagnum_teres,    clade2, Sphagn~rrosum,    clade3 
#> min values  : 0.0000000,  6.170776e-103, 0.0000000, 8.557791e-170, 0.0000000 
#> max values  : 0.9318513,   2.934258e-01, 0.9318242,  3.475139e-01, 0.9198755